# API-Go/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  auth.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  response.Body:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'API REST de Identity and Access Management (IAM): autenticación JWT,
    RBAC, usuarios, roles, funcionalidades, sesiones y auditoría.'
  title: API JeussAirel (IAM)
  version: "1.0"
paths:
  /audit/logins:
    get:
      description: Devuelve la auditoría de inicios de sesión paginada.
      parameters:
      - description: Número de página (default 1)
        in: query
        name: page
        type: integer
      - description: Tamaño de página (1-100, default 20)
        in: query
        name: page_size
        type: integer
      responses:
        "200":
          description: Auditoría obtenida
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Acceso denegado (AUDIT_LOG_VIEW:read)
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Listar auditoría de logins
      tags:
      - Audit
  /audit/logins/users/{userId}:
    get:
      description: Devuelve el historial de inicios de sesión de un usuario.
      parameters:
      - description: UUID del usuario
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: Auditoría del usuario
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Acceso denegado (AUDIT_LOG_VIEW:read)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Usuario no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Auditoría de login de un usuario
      tags:
      - Audit
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Genera una solicitud de reseteo. Si el email existe, se envía la
        instrucción (token de recuperación).
      parameters:
      - description: Email del usuario
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Si el email existe, se envió la instrucción
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/response.Body'
      summary: Solicitar recuperación de contraseña
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica al usuario con email y contraseña y devuelve el par de
        tokens (access + refresh).
      parameters:
      - description: Credenciales de acceso
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Inicio de sesión exitoso
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/auth.TokenResponse'
              type: object
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Credenciales inválidas
          schema:
            $ref: '#/definitions/response.Body'
      summary: Iniciar sesión
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoca el refresh token y cierra la sesión activa.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sesión cerrada
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Cerrar sesión
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Renueva el access token (y el refresh token) usando un refresh
        token válido.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens renovados
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/auth.TokenResponse'
              type: object
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Refresh token inválido o expirado
          schema:
            $ref: '#/definitions/response.Body'
      summary: Renovar tokens
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Cambia la contraseña usando el token de recuperación generado en
        forgot-password.
      parameters:
      - description: Token y nueva contraseña
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Contraseña actualizada
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Token inválido o expirado
          schema:
            $ref: '#/definitions/response.Body'
      summary: Restablecer contraseña
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT de acceso. Uso: `Authorization: Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"